edit code from reivew
