+++ Gemfile
@@ -17,6 +17,7 @@ gem "omniauth-facebook"
 gem 'rubocop-rails', require: false
 gem "rubyzip"
 gem 'font-awesome-sass', '~> 5.12.0'
+gem "sidekiq"
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '~> 6.0.2', '>= 6.0.2.1'
