+++ app/jobs/notifications_to_email_job.rb
@@ -1,0 +1,7 @@
+class NotificationsToEmailJob < ApplicationJob
+  queue_as :default
+
+  def perform(user, content)
+    NotificationsMailer.notifications_from_app(user, content).deliver_now
+  end
+end
