+++ test/jobs/notifications_to_email_job_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class NotificationsToEmailJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
